{add} Actualizando el seeder run y el factorie de post

+++ sites/eloquent/database/migrations/2021_03_13_020425_create_posts_table.php
@@ -15,7 +15,7 @@ class CreatePostsTable extends Migration
     {
         Schema::create('posts', function (Blueprint $table) {
             $table->id();
-            $table->Integer('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('title');
 
             $table->foreign('user_id')->references('id')->on('users');
